Harden documentation example tests
Accept CommonMark's permitted fence indentation while retaining strict
marker and closing-fence validation.

Run documented Cargo workflows and binaries behind closed environment
allow-lists and temporary target directories. Add executable probes so
future changes cannot silently reintroduce inherited host state.

Document the test-helper contracts and remove local recovery metadata from
the published execution plan.
