+++ ortho_config/tests/documentation_examples_rust_tests.rs
@@ -5,8 +5,8 @@ mod documentation_examples;
 mod workspace;
 
 use anyhow::{Context, Result, ensure};
-use documentation_examples::documented_example;
-use std::path::Path;
+use documentation_examples::{DocumentedExample, documented_example};
+use std::path::{Path, PathBuf};
 use workspace::{DependencyAlias, EnvironmentVariable, ExampleId, ExampleWorkspace, RunFile};
 
 const STANDARD_RUST_EXAMPLES: &[&str] = &[
@@ -122,6 +122,53 @@ fn workspace_rejects_paths_outside_an_example_directory() -> Result<()> {
         );
     }
 
+    Ok(())
+}
+
+#[test]
+fn workspace_sanitizes_the_documented_binary_environment() -> Result<()> {
+    ensure!(
+        std::env::var_os("PATH").is_some(),
+        "the parent test process should provide PATH"
+    );
+    let workspace = ExampleWorkspace::new(DependencyAlias("ortho_config"))?;
+    workspace.add_binary(&DocumentedExample {
+        id: "environment-probe".to_owned(),
+        language: "rust".to_owned(),
+        body: concat!(
+            "fn main() {\n",
+            "    let current = std::env::current_dir().expect(\"read current directory\");\n",
+            "    let home = std::env::var_os(\"HOME\").expect(\"HOME should be set\");\n",
+            "    let xdg = std::env::var_os(\"XDG_CONFIG_HOME\").expect(\"XDG should be set\");\n",
+            "    println!(\"current={}\", current.display());\n",
+            "    println!(\"home={}\", std::path::Path::new(&home).display());\n",
+            "    println!(\"xdg={}\", std::path::Path::new(&xdg).display());\n",
+            "    println!(\"path_present={}\", std::env::var_os(\"PATH\").is_some());\n",
+            "}\n",
+        )
+        .to_owned(),
+        source: "environment probe",
+        line: 1,
+    })?;
+    workspace.build()?;
+
+    let output = workspace.run(ExampleId("environment-probe"), std::iter::empty::<&str>())?;
+    ensure!(output.status.success(), "environment probe should succeed");
+    let stdout = String::from_utf8(output.stdout).context("environment probe output is UTF-8")?;
+    let values = stdout
+        .lines()
+        .filter_map(|line| line.split_once('='))
+        .collect::<std::collections::HashMap<_, _>>();
+    let value = |name| {
+        values
+            .get(name)
+            .copied()
+            .with_context(|| format!("environment probe should report {name}"))
+    };
+    let current = PathBuf::from(value("current")?);
+    ensure!(Path::new(value("home")?) == current);
+    ensure!(Path::new(value("xdg")?) == current.join("xdg"));
+    ensure!(value("path_present")? == "false");
     Ok(())
 }
